Improve prepending url_helpers of RouteSet
